// src/api/models/Project.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamBoard API
 * Kanban Board API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { User } from './User';
import {
    UserFromJSON,
    UserFromJSONTyped,
    UserToJSON,
    UserToJSONTyped,
} from './User';
import type { Column } from './Column';
import {
    ColumnFromJSON,
    ColumnFromJSONTyped,
    ColumnToJSON,
    ColumnToJSONTyped,
} from './Column';

/**
 * 
 * @export
 * @interface Project
 */
export interface Project {
    /**
     * 
     * @type {number}
     * @memberof Project
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof Project
     */
    description?: string;
    /**
     * 
     * @type {Array<User>}
     * @memberof Project
     */
    members?: Array<User>;
    /**
     * 
     * @type {Array<Column>}
     * @memberof Project
     */
    columns?: Array<Column>;
    /**
     * 
     * @type {User}
     * @memberof Project
     */
    createdBy?: User;
    /**
     * 
     * @type {Date}
     * @memberof Project
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Project
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the Project interface.
 */
export function instanceOfProject(value: object): value is Project {
    return true;
}

export function ProjectFromJSON(json: any): Project {
    return ProjectFromJSONTyped(json, false);
}

export function ProjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): Project {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'members': json['members'] == null ? undefined : ((json['members'] as Array<any>).map(UserFromJSON)),
        'columns': json['columns'] == null ? undefined : ((json['columns'] as Array<any>).map(ColumnFromJSON)),
        'createdBy': json['createdBy'] == null ? undefined : UserFromJSON(json['createdBy']),
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function ProjectToJSON(json: any): Project {
    return ProjectToJSONTyped(json, false);
}

export function ProjectToJSONTyped(value?: Project | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'members': value['members'] == null ? undefined : ((value['members'] as Array<any>).map(UserToJSON)),
        'columns': value['columns'] == null ? undefined : ((value['columns'] as Array<any>).map(ColumnToJSON)),
        'createdBy': UserToJSON(value['createdBy']),
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
    };
}